Fix KTS generation
Plugin version in new plugin block used the wrong string quote
character in KTS (' instead of "). This was due to a lack for tests,
so this change also adds new tests with pure java-library and pure
kotlin projects.

This change also does a bit of clean-up to avoid creating the
buildscript and plugins blocks in the root project if they are
not needed.